refactor: standardize code formatting and improve readability
@ kukac

+++ Au-O/components/Post/base/PostOptionMenu.tsx
@@ -1,19 +1,19 @@
-import React from "react";
-import { PostCreationTexts, PostStatusTexts } from "@/constants/texts";
-import { ActionSheetIOS, Alert, Modal, Platform, View } from "react-native";
-import * as Clipboard from "expo-clipboard";
-import Toast from "react-native-toast-message";
-import { Picker } from "@react-native-picker/picker";
-import { router } from "expo-router";
-import { Colors } from "@/constants/Colors";
-import { apiFetch, deleteImgurImage } from "@/lib/apiClient";
+import React from 'react';
+import { PostCreationTexts, PostStatusTexts } from '@/constants/texts';
+import { ActionSheetIOS, Alert, Modal, Platform, View } from 'react-native';
+import * as Clipboard from 'expo-clipboard';
+import Toast from 'react-native-toast-message';
+import { Picker } from '@react-native-picker/picker';
+import { router } from 'expo-router';
+import { Colors } from '@/constants/Colors';
+import { apiFetch, deleteImgurImage } from '@/lib/apiClient';
 export default function PostOptionMenu(
   preview: boolean,
-  language: "EN" | "HU",
+  language: 'EN' | 'HU',
   postId: number,
   userId: number | null,
   authorId: number | null,
-  onDelete?: () => void,
+  onDelete?: () => void
 ) {
   if (preview) return null;
   let iosOptions = [
@@ -27,52 +27,51 @@ export default function PostOptionMenu(
   async function handleShare() {
     await Clipboard.setStringAsync(postId.toString());
     Toast.show({
-      type: "success",
+      type: 'success',
       text1: PostCreationTexts.copied[language],
-      position: "top",
+      position: 'top',
       visibilityTime: 2000,
     });
   }
   async function handleReport() {
-    console.log("report");
-
+    console.log('report');
   }
   async function handleEdit() {
     if (authorId && userId && authorId === userId) {
-      router.push({pathname: "/(post)/edit/[id]", params: {id: postId.toString()}});
+      router.push({ pathname: '/(post)/edit/[id]', params: { id: postId.toString() } });
     }
   }
   async function handleDelete() {
     if (authorId && userId && authorId === userId) {
-      Alert.alert("Post deletion", "Are you sure you want to delete this post?", [{
-        text: PostStatusTexts.deletePrompt.buttons.cancel[language],
-        style: "cancel",
-      }, {
-        text: PostStatusTexts.deletePrompt.buttons.delete[language],
-        style: "destructive",
-        onPress: async () => {
-          
-           const res = await apiFetch(`posts/post/${postId}`, "DELETE", true);
-          if (res && res.status === 200) {
-            Toast.show({
-              type: "success",
-              text1: PostStatusTexts.deletePrompt.success[language],
-            });
-            if (onDelete) onDelete();
-          }
-          else {
-            Toast.show({
-              type: "error",
-              text1: PostStatusTexts.deletePrompt.error[language],
-            });
-          }
-         
-        }
-      }])
+      Alert.alert('Post deletion', 'Are you sure you want to delete this post?', [
+        {
+          text: PostStatusTexts.deletePrompt.buttons.cancel[language],
+          style: 'cancel',
+        },
+        {
+          text: PostStatusTexts.deletePrompt.buttons.delete[language],
+          style: 'destructive',
+          onPress: async () => {
+            const res = await apiFetch(`posts/post/${postId}`, 'DELETE', true);
+            if (res && res.status === 200) {
+              Toast.show({
+                type: 'success',
+                text1: PostStatusTexts.deletePrompt.success[language],
+              });
+              if (onDelete) onDelete();
+            } else {
+              Toast.show({
+                type: 'error',
+                text1: PostStatusTexts.deletePrompt.error[language],
+              });
+            }
+          },
+        },
+      ]);
     }
   }
 
-  if (Platform.OS === "ios") {
+  if (Platform.OS === 'ios') {
     return ActionSheetIOS.showActionSheetWithOptions(
       {
         options: iosOptions,
@@ -96,7 +95,7 @@ export default function PostOptionMenu(
         }
       }
     );
-  } else if (Platform.OS === "android") {
+  } else if (Platform.OS === 'android') {
     return (
       <Picker
         onValueChange={async (itemIndex) => {
@@ -122,8 +121,8 @@ export default function PostOptionMenu(
         <Picker.Item label={PostCreationTexts.options.report[language]} />
         {userId && userId === authorId && (
           <>
-          <Picker.Item color={Colors.highlight.main} label={PostCreationTexts.options.edit[language]} />
-          <Picker.Item color={Colors.highlight.main} label={PostCreationTexts.deletePost[language]} />
+            <Picker.Item color={Colors.highlight.main} label={PostCreationTexts.options.edit[language]} />
+            <Picker.Item color={Colors.highlight.main} label={PostCreationTexts.deletePost[language]} />
           </>
         )}
       </Picker>
